+++ core/config.py
@@ -91,6 +91,7 @@ class ConfigManager:
         "silent_alert_on_mention": False,
         "show_timestamp": True,
         "anonymous_snippets": False,
+        "plain_snippets": False,
         "require_close_reason": False,
         # group conversations
         "private_added_to_group_title": "New Thread (Group)",
@@ -207,6 +208,7 @@ class ConfigManager:
         "update_notifications",
         "thread_contact_silently",
         "anonymous_snippets",
+        "plain_snippets",
         "require_close_reason",
         "recipient_thread_close",
         "thread_show_roles",